refactor: Product 컴포넌트 및 스타일 분리 Co-authored-by: youdame <112458620+youdame@users.noreply.github.com>
